+++ R/zzz.R
@@ -10,6 +10,8 @@
 #'     the julia in path.
 #' @param verbose whether to print out detailed information
 #'     about \code{julia_setup}.
+#' @param install whether to execute install script for dependent julia packages, whose default value is TRUE;
+#'     but can be set to FALSE to save startup time when no installation of dependent julia packages is needed.
 #' @param force whether to force julia_setup to execute again.
 #' @param useRCall whether or not you want to use RCall.jl in julia,
 #'     which is an amazing package to access R in julia.
@@ -24,7 +26,7 @@
 #' }
 #'
 #' @export
-julia_setup <- function(JULIA_HOME = NULL, verbose = TRUE, force = FALSE, useRCall = TRUE) {
+julia_setup <- function(JULIA_HOME = NULL, verbose = TRUE, install = TRUE, force = FALSE, useRCall = TRUE) {
     ## libR <- paste0(R.home(), '/lib')
     ## system(paste0('export LD_LIBRARY_PATH=', libR, ':$LD_LIBRARY_PATH'))
 
@@ -67,13 +69,17 @@ julia_setup <- function(JULIA_HOME = NULL, verbose = TRUE, force = FALSE, useRCa
     ## if (verbose) message("Julia initiation...")
 
     if (.Platform$OS.type == "windows") {
-        libm <- julia_line(c("-e", "print(Libdl.dlpath(Base.libm_name))"), stdout = TRUE)
-        dyn.load(libm, DLLpath = .julia$bin_dir)
+        ## libm is needed to load seperately only for Julia version 0.6.x
 
-        # following is required to load dll dependencies from JULIA_HOME
-        cur_dir <- getwd()
-        setwd(.julia$bin_dir)
-        on.exit(setwd(cur_dir))
+        if (newer("0.6.5", .julia$VERSION)) {
+            libm <- julia_line(c("-e", "print(Libdl.dlpath(Base.libm_name))"), stdout = TRUE)
+            dyn.load(libm, DLLpath = .julia$bin_dir)
+
+            # following is required to load dll dependencies from JULIA_HOME
+            cur_dir <- getwd()
+            setwd(.julia$bin_dir)
+            on.exit(setwd(cur_dir))
+        }
     }
 
     juliacall_initialize(.julia$dll_file)
@@ -102,7 +108,9 @@ julia_setup <- function(JULIA_HOME = NULL, verbose = TRUE, force = FALSE, useRCa
 
     if (verbose) message("Loading setup script for JuliaCall...")
 
-    install_dependency()
+    if (isTRUE(install)) {
+        install_dependency()
+    }
 
     if (!newer(.julia$VERSION, "0.7.0")) {
         ## message("Before 0.7.0")